Move ObservableSplit into internals
